added cli for resetting elasticsearch indexes

+++ morpfw/cli.py
@@ -28,7 +28,7 @@ from alembic.config import main as alembic_main
 from .alembic import drop_all
 from .main import create_admin, default_settings
 from .request import request_factory
-
+from random import randint
 
 def load_settings(settings_file, default=default_settings):
     dsettings = os.environ.get("MORP_SETTINGS", {})
@@ -90,6 +90,20 @@ def load(settings_file: str = None, host: str = None, port: int = None):
         "port": port,
     }
 
+
+def confirmation_dialog(message='Are you sure you want to proceed?',
+        token_length=6):
+
+    token = ''
+    for i in range(token_length):
+        x = randint(0,9)
+        token += str(x)
+
+    ans = input("%s \n(Please enter '%s' to confirm): " % (message, token))
+    if ans.strip() != token:
+        print('Invalid confirmation token')
+        return False
+    return True
 
 @click.group()
 @click.option("-s", "--settings", type=str, default=None, help="path to settings.yml")
@@ -291,6 +305,9 @@ def resetdb(ctx):
 
     settings = param["settings"]
 
+    if not confirmation_dialog():
+        return 
+
     with request_factory(settings) as request:
         drop_all(request)
 
@@ -340,6 +357,31 @@ def create_esindex(ctx):
                     print("OK")
                 else:
                     print("EXIST")
+
+
+@cli.command(help="delete all elasticsearch indexes")
+@click.pass_context
+def reset_esindex(ctx):
+
+    if not confirmation_dialog():
+        return 
+
+    param = load(ctx.obj["settings"])
+
+    settings = param["settings"]
+
+    with request_factory(settings) as request:
+
+        types = request.app.config.type_registry.get_typeinfos(request)
+        client = request.get_es_client()
+        for typeinfo in types.values():
+            collection = request.get_collection(typeinfo["name"])
+            storage = collection.storage
+            if isinstance(storage, morpfw.ElasticSearchStorage):
+                print("Deleting index %s .. " % storage.index_name, end="")
+                client.indices.delete(storage.index_name)
+                print("OK")
+
 
 
 def main():
